Moved/renamed 'type' package to 'model' package and updated imports and references
Conflicts:

	src/soaplib/model/primitive.py
	src/soaplib/test/test_application.py
