+++ doc/source/pages/message_api.rst
@@ -8,7 +8,7 @@ programatically. The Message object follows the to_xml/from_xml usage pattern as
 the type. ::
 
     >>> from soaplib.soap import Message
-    >>> from soaplib.type.primitive import *
+    >>> from soaplib.model.primitive import *
     >>> import cElementTree as et
     >>> message = Message('myFunction',[('a',String),('b',Integer),('c',Float)])
     >>> print et.tostring(message.to_xml('a',13,3.14))
