+++ examples/async.py
@@ -22,9 +22,9 @@ from threading import Thread
 
 
 from soaplib.service import rpc, DefinitionBase
-from soaplib.type.primitive import String, Integer
+from soaplib.model.primitive import String, Integer
 from soaplib.util import get_callback_info
-from soaplib.pattern.server.wsgi import Application
+from soaplib.server.wsgi import Application
 
 
 '''
