+++ src/soaplib/client/_base.py
@@ -1,9 +1,9 @@
 
-from soaplib.type.exception import Fault
+from soaplib.model.exception import Fault
 
 from lxml import etree
 from soaplib import MethodContext
-from soaplib.type.primitive import string_encoding
+from soaplib.model.primitive import string_encoding
 
 class Factory(object):
     def __init__(self, app):
