+++ src/soaplib/model/base.py
