+++ src/soaplib/model/enum.py
