+++ src/soaplib/server/wsgi.py
@@ -29,8 +29,8 @@ import traceback
 
 import soaplib
 
-from soaplib.type.exception import Fault
-from soaplib.type.primitive import string_encoding
+from soaplib.model.exception import Fault
+from soaplib.model.primitive import string_encoding
 
 from soaplib.mime import apply_mtom
 from soaplib.mime import collapse_swa
