+++ src/soaplib/test/interop/_test_client_base.py
@@ -19,7 +19,7 @@
 
 import unittest
 
-from soaplib.type.exception import Fault
+from soaplib.model.exception import Fault
 
 from datetime import datetime
