+++ src/soaplib/test/test_soap.py
@@ -21,15 +21,15 @@ import unittest
 
 from lxml import etree
 
-from soaplib.type.clazz import ClassSerializer
+from soaplib.model.clazz import ClassSerializer
 
-from soaplib.type.clazz import Array
-from soaplib.type.primitive import DateTime
-from soaplib.type.primitive import Float
-from soaplib.type.primitive import Integer
-from soaplib.type.primitive import String
+from soaplib.model.clazz import Array
+from soaplib.model.primitive import DateTime
+from soaplib.model.primitive import Float
+from soaplib.model.primitive import Integer
+from soaplib.model.primitive import String
 
-from soaplib.type.clazz import ClassSerializer as Message
+from soaplib.model.clazz import ClassSerializer as Message
 from soaplib import from_soap
 
 class Address(ClassSerializer):
