+++ src/soaplib/test/type/test_clazz.py
@@ -20,13 +20,13 @@
 import datetime
 import unittest
 
-from soaplib.type.clazz import ClassSerializer
-from soaplib.type.clazz import Array
+from soaplib.model.clazz import ClassSerializer
+from soaplib.model.clazz import Array
 
-from soaplib.type.primitive import DateTime
-from soaplib.type.primitive import Float
-from soaplib.type.primitive import Integer
-from soaplib.type.primitive import String
+from soaplib.model.primitive import DateTime
+from soaplib.model.primitive import Float
+from soaplib.model.primitive import Integer
+from soaplib.model.primitive import String
 
 from lxml import etree
